clean compile warning (#2380)

+++ core/src/segment/Attr.cpp
@@ -132,7 +132,7 @@ Attr::Erase(std::vector<int32_t>& offsets) {
     auto loop_size = uids_.size();
 
     for (size_t i = 0; i < loop_size;) {
-        while (skip != offsets.cend() && i == *skip) {
+        while (skip != offsets.cend() && i == (size_t)(*skip)) {
             ++i;
             ++skip;
         }
